Functioning calculator
I am disappointed in myself for how long it took me to make this first
commit. I have so far made the HTML/CSS for a calculator, as well as 90%
of functionality. I have yet to do the Undo button, as well as some JS
to change operator button colors when pressed. I also will be adding in
functionality to input decimals, as well as an error message for someone
trying to divide by zero. Commits will be more frequent from here on
out, I have been so focused on trying to figure out continuous
operations and variable storage that I forgot to commit.