Design improvements, added undo functionality
Added "Made in Mexico" to the bottom of the calculator to reflect the
work I have done for this class while competing in professional baseball
in Mexico. Added small shadows to the buttons, to make them look more
natural with the shadowing of the calculator. Added undo button
functionality, where you can now backspace if you typed in the wrong
number.