Added handling of decimals
installed a function to evaluate the answer to my operate function, and
if the answer is not an integer, it rounds the answer to four decimal
places. I found this to be the least confusing, as before many answers
showed as 'x.00' which made the user think the answer was an integer
when it was not.