Added comments/cleaning and operator coloring
Added comments to make code more readable in the future, as well as
cleaned up a line of code to use function instead of assigning variable.

Changed rounding to two decimals instead of four, feel as though it's
cleaner and makes sense in more situations.

Calculator now has operator button coloring functionality. Super
excited about that, as it was semi difficult to figure out, but was my
idea to add and not suggested by the course. Should really improve user
experience.